chore: refactor ui so we can render create pages
